tidy up FpsTimer
